Move ItemsAdminActivity.kt and ItemRequestActivity.kt to correct folders
