+++ app/src/main/java/com/github/sdpteam15/polyevents/view/fragments/home/AdminHomeFragment.kt
@@ -9,7 +9,7 @@ import android.widget.Button
 import androidx.fragment.app.Fragment
 import com.github.sdpteam15.polyevents.R
 import com.github.sdpteam15.polyevents.model.entity.UserRole
-import com.github.sdpteam15.polyevents.view.activity.ItemsAdminActivity
+import com.github.sdpteam15.polyevents.view.activity.admin.ItemsAdminActivity
 import com.github.sdpteam15.polyevents.view.activity.MainActivity
 import com.github.sdpteam15.polyevents.view.activity.admin.*
