+++ app/src/main/java/com/github/sdpteam15/polyevents/view/fragments/home/ProviderHomeFragment.kt
@@ -2,7 +2,6 @@ package com.github.sdpteam15.polyevents.view.fragments.home
 
 import android.content.Intent
 import android.os.Bundle
-import android.provider.ContactsContract
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,7 +10,7 @@ import androidx.fragment.app.Fragment
 import com.github.sdpteam15.polyevents.R
 import com.github.sdpteam15.polyevents.model.database.remote.Database
 import com.github.sdpteam15.polyevents.model.entity.UserRole
-import com.github.sdpteam15.polyevents.view.activity.ItemRequestActivity
+import com.github.sdpteam15.polyevents.view.activity.activityprovider.ItemRequestActivity
 import com.github.sdpteam15.polyevents.view.activity.MainActivity
 import com.github.sdpteam15.polyevents.view.activity.activityprovider.MyItemRequestsActivity
 import com.github.sdpteam15.polyevents.view.activity.activityprovider.EventManagementActivityProvider
@@ -25,7 +24,7 @@ class ProviderHomeFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         val viewRoot = inflater.inflate(R.layout.fragment_home_provider, container, false)
         viewRoot.findViewById<Button>(R.id.id_request_button).setOnClickListener {
             val intent = Intent(activity, ItemRequestActivity::class.java)
