+++ app/src/main/res/layout/activity_item_request.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.activity.ItemRequestActivity">
+    tools:context=".view.activity.activityprovider.ItemRequestActivity">
 
     <LinearLayout
         android:id="@+id/layout_spinner"
