+++ app/src/main/res/layout/activity_items_admin.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.activity.ItemsAdminActivity">
+    tools:context=".view.activity.admin.ItemsAdminActivity">
 
     <ImageButton
         android:id="@+id/id_add_item_button"
